+++ .vuepress/config.js
@@ -106,8 +106,11 @@ module.exports = {
             text: '后端：微服务架构 Kubernetes',
             link: 'https://github.com/fenixsoft/microservice_arch_kubernetes'
           }, {
-            text: '后端：无服务架构 Knative',
-            link: 'https://github.com/fenixsoft/serverless_arch_knative'
+            text: '后端：微服务架构 Istio',
+            link: 'https://github.com/fenixsoft/microservice_arch_istio'
+          }, {
+            text: '后端：无服务架构 Serverless',
+            link: 'https://github.com/fenixsoft/serverless_arch'
           }
         ]
       },
@@ -147,7 +150,8 @@ module.exports = {
               '/exploration/projects/monolithic_arch_springboot',
               '/exploration/projects/microservice_arch_springcloud',
               '/exploration/projects/microservice_arch_kubernetes',
-              '/exploration/projects/serverless_arch_knative'
+              '/exploration/projects/microservice_arch_istio',
+              '/exploration/projects/serverless_arch'
             ]
           }]
       }, {